fix: address Codex review feedback on scheduler skip path
- Move the KV-fairness regression test into the existing
  tests/v1/core/test_scheduler.py suite instead of a new one-test
  file, per AGENTS.md's guidance to extend a nearby suite.
- Track allocation-failure skips separately (waiting_capacity_bound)
  and OR it into prefill_capacity_bound. A capacity-blocked request
  moves from self.waiting into step_skipped_waiting, so
  bool(self.waiting) alone can go False even though the step was
  genuinely capacity-bound, incorrectly re-arming DP prefill
  throttling's defer_prefills on the next throttled step.
- Add test_throttle_capacity_bound_guard_admits_after_kv_blocked_skip,
  which fails without the fix (verified by reverting it locally) and
  passes with it, covering the KV-block-allocation-failure path that
  the existing token-budget-only capacity-bound test didn't exercise.
