--- tests/v1/core/test_scheduler_kv_fairness.py
@@ -1,28 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-
-from tests.v1.core.utils import create_requests, create_scheduler
-
-
-def test_waiting_kv_blocked_request_does_not_block_lighter_request():
-    scheduler = create_scheduler(
-        max_num_batched_tokens=8,
-        max_num_seqs=2,
-        num_blocks=2,
-        block_size=4,
-        enable_chunked_prefill=True,
-    )
-
-    heavy = create_requests(
-        num_requests=1, num_tokens=9, block_size=4, req_ids=["heavy"]
-    )[0]
-    light = create_requests(
-        num_requests=1, num_tokens=4, block_size=4, req_ids=["light"]
-    )[0]
-
-    scheduler.add_request(heavy)
-    scheduler.add_request(light)
-
-    output = scheduler.schedule()
-    scheduled_ids = [req.req_id for req in output.scheduled_new_reqs]
-    assert scheduled_ids == ["light"]
